Keep the artwork surface fractional, so a 5mm box is drawn 5mm wide
The viewport handed to Batik is fractional, but the surface the artwork is
painted into was truncated to whole pixels: the form XObject's BBox in the
PDF device, and the clip and origin of the Graphics2D in the Java2D one. So
artwork sized in units that do not land on a pixel, such as 5mm, lost its
last sliver.

Four goldens gain that sliver along the top edge, and the 5mm bar in
SvgCssImageNonVisualTest now covers the nineteenth, nine tenths covered,
pixel column rather than stopping at 18.

+++ openhtmltopdf-pdfbox/src/main/java/com/openhtmltopdf/pdfboxout/PdfBoxFastOutputDevice.java
@@ -1187,7 +1187,10 @@ public class PdfBoxFastOutputDevice extends AbstractOutputDevice implements Outp
 
     private PDFormXObject createXForm(float width, float height, OutputDeviceGraphicsDrawer renderer) {
         try {
-            PdfBoxGraphics2D pdfBoxGraphics2D = new PdfBoxGraphics2D(_writer, (int) width, (int) height);
+            // The size given here is the BBox of the form, so anything drawn outside it is
+            // clipped away. Keep it fractional: truncating to whole pixels shaves the last
+            // sliver off artwork sized in units that do not land on a pixel, such as 5mm.
+            PdfBoxGraphics2D pdfBoxGraphics2D = new PdfBoxGraphics2D(_writer, width, height);
 			/*
 			 * Create and set the fontTextDrawer to perform the font mapping.
 			 */
